+++ test/integration/resource_group.rb
@@ -14,26 +14,38 @@ resource = Fog::Resources::AzureRM.new(
   client_secret: azure_credentials['client_secret'],
   subscription_id: azure_credentials['subscription_id']
 )
-
+resource_group_name = "RG-#{current_time}"
 ########################################################################################################################
 ######################                    Create Resource Group                                   ######################
 ########################################################################################################################
 
-resource.resource_groups.create(
-  name: 'TestRG-ZN',
-  location: LOCATION
-)
+begin
+  resource_group = resource.resource_groups.create(
+    name: resource_group_name,
+    location: 'eastus'
+  )
 
-########################################################################################################################
-######################           Get All Resource Groups in a Subscription                        ######################
-########################################################################################################################
+  puts "Created resource group #{resource_group.name}"
+  ########################################################################################################################
+  ######################           Get All Resource Groups in a Subscription                        ######################
+  ########################################################################################################################
 
-resource_groups = resource.resource_groups
+  resource_groups = resource.resource_groups
+  puts 'List resource groups:'
+  resource_groups.each do |a_resource_group|
+    puts a_resource_group.name
+  end
 
-resource_group = resource_groups.get('TestRG-ZN')
+  resource_group = resource_groups.get(resource_group_name)
+  puts "Get resource group: #{resource_group.name}"
 
-########################################################################################################################
-######################                           Destroy Resource Group                           ######################
-########################################################################################################################
+  ########################################################################################################################
+  ######################                           Destroy Resource Group                           ######################
+  ########################################################################################################################
 
-resource_group.destroy
+  puts "Deleted resource group: #{resource_group.destroy}"
+
+rescue
+  puts 'Integration Test for resource group is failing'
+  resource_group.destroy unless resource_group.nil?
+end
